iOS Ping functionality (#602)
* Added ping button and rearranged traceroute button

* Progress on ping. Refactored some traceroute ui config to be reusued.

* Started working on adding the ping locations view

* Progress on ping functionality

* Automatically start ping after selection

* Cleanup

* Improve delegate method name

* Remove target-specific deployment target and update project deployment target

* Update credits.html

* Remove unnecessary availability check and remove some unnecessary frame calculations

* Remove CarPlay, Mac, and Apple Watch icon entries

* Expose immutable array of packet records

+++ External Code/SCPinions/SCPacketUtility/SCPacketRecord.h
@@ -34,7 +34,7 @@
 @interface SCPacketRecord : NSObject
 
 @property uint16_t sequenceNumber; //A sequence number for tracking of packet relationships as we perform network operations
-@property int sentWithTTL; // The TTL that our original packet has assigned to it
+@property NSInteger sentWithTTL; // The TTL that our original packet has assigned to it
 @property NSString* responseAddress; //The address of the machine that sent us a respponse packet for this sequence number
 @property NSDate* departure; // Departure time of packet sent for this sequence number
 @property NSDate* arrival; // Arrive of response packet for this sequence number
